+++ frontend/src/pages/Search.tsx
@@ -3,22 +3,21 @@ import { Loading } from "../components/Loading";
 import { Pagination } from "../components/Pagination";
 import { SelectBox } from "../components/SelectBox";
 import { useRoom } from "../hooks/useAxios";
-import { Gallery } from "../layouts/listing/Gallery";
+import { Gallery } from "../layouts/listing/gallery/index";
 import { ListItem } from "../types/room.type";
 
 export const Search = () => {
-  const { data, isError, isLoading } = useRoom({ method: "GET" });
+  const { data, isError, isLoading } = useRoom();
 
   if (isError)
     return (
       <ErrorMsg msg="Sorry, something wrong with the connection." isReloadBtn={true} />
     );
   if (isLoading) return <Loading />;
-
-  const rooms: ListItem[] = data.rooms;
-  if (rooms.length === 0)
+  if (!data || data.rooms.length === 0)
     return <ErrorMsg msg="Sorry, no rooms found." isReloadBtn={false} />;
 
+  const rooms: ListItem[] = data.rooms;
   const galleries = rooms.map((room) => <Gallery key={room.id} {...room} />);
 
   return (
